### 2022-12-26 #5018 @jbytecode
@mitchellpound - while you are preparing a suggested list of suitable reviewers, It is the right time to take a look at the paper. Please consider correcting/adding things mentioned below:

- The paper does not cite any previous documents. This is not usual, it is a sign of a missing literature review, otherwise it is assumed that you find all of the things mentioned in the manuscript which is not possible.
- Please take a look at the accepted JOSS submissions and have an idea to refactor the manuscript later. The manuscript and the software will be reshaped during the review process so you don't need to refactor the text now.
