### 2022-12-29 #5018 @mitchellpound
@jbytecode,
Thank you for being willing to be the handling editor.

I would suggest the following reviewers from the list based off their knowledge of finance as well as Julia.
- taqtiqa-mark
- lrnv
- daviddewhurst
- AvianaGlobal
