### 2022-12-29 #5018 @mitchellpound
@jbytecode - Thank you for letting me know about the citations. We included several sources that were key to developing the Bruno package in the paper.bib file. However, it doesn't seem to be linking to the paper when it is generated from the paper.md file. I tried to copy the syntax from the JOSS documentation, but must be missing something. Do you have any suggestions on how to get the bibliography to link and generate with the rest of the paper?
