### 2022-12-29 32a4218
add missing DOIs
