+++ paper/paper.bib
@@ -13,7 +13,8 @@
 	publisher={John Wiley & Sons Incorporated}, 
 	author={Clewlow, Les and Strickland, Chris}, 
 	year={1998}, 
-	month={Jun} 
+	month={Jun},
+  doi={10.1057/9780230392687.0018} 
 }
 
 @article{politis1994stationary,
@@ -35,7 +36,8 @@
   number={1},
   pages={53--70},
   year={2004},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi={10.1081/etc-120028836}
 }
 
 @article{patton2009correction,
@@ -46,7 +48,8 @@
   number={4},
   pages={372--375},
   year={2009},
-  publisher={Taylor \& Francis}
+  publisher={Taylor \& Francis},
+  doi={10.1080/07474930802459016}
 }
 
 @misc{pound_2022, 
