### 2022-12-29 afc9ba7
add citation for Julia in bibtex
