+++ paper/paper.bib
@@ -1,3 +1,15 @@
+ @article{bezanson2017julia,
+  title={Julia: A fresh approach to numerical computing},
+  author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
+  journal={SIAM Review},
+  volume={59},
+  number={1},
+  pages={65--98},
+  year={2017},
+  publisher={SIAM},
+  doi={10.1137/141000671}
+}
+
  @book{mcdonald_2013, 
  	place={Boston}, 
 	title={Derivatives markets}, 
