### 2022-12-29 #5018 @jbytecode
@mitchellpound - I've just sent a pull requests that adds some missing DOIs. The PR also cites Julia in paper markdown to show how to link citation entries of bibtex files from the main text. Please link the required citations in the manuscript as in the PR and ping me when you are ready to jump to the next step. Thank you in advance.

[The PR](https://github.com/USU-Analytics-Solution-Center/Bruno.jl/pull/59)
