### 2022-12-30 356d876
Added Culp book citation to paper.md and .bib files
