+++ paper/paper.bib
@@ -64,6 +64,13 @@
   doi={10.1080/07474930802459016}
 }
 
+@book{culp2011risk,
+  title={Risk transfer: derivatives in theory and practice},
+  author={Culp, Christopher L},
+  year={2011},
+  publisher={John Wiley \& Sons}
+}
+
 @misc{pound_2022, 
 	place={Utah State University}, 
 	title={Black Scholes Delta Hedge in Imperfect Markets}, 
