### 2022-12-30 66e1bb4
Added references in text to paper.md
