+++ paper/paper.bib
@@ -10,6 +10,17 @@
   doi={10.1137/141000671}
 }
 
+@article{black1973pricing,
+  title={The pricing of options and corporate liabilities},
+  author={Black, Fischer and Scholes, Myron},
+  journal={Journal of political economy},
+  volume={81},
+  number={3},
+  pages={637--654},
+  year={1973},
+  publisher={The University of Chicago Press}
+}
+
  @book{mcdonald_2013, 
  	place={Boston}, 
 	title={Derivatives markets}, 
@@ -69,12 +80,4 @@
   author={Culp, Christopher L},
   year={2011},
   publisher={John Wiley \& Sons}
-}
-
-@misc{pound_2022, 
-	place={Utah State University}, 
-	title={Black Scholes Delta Hedge in Imperfect Markets}, 
-	author={Pound, Mitchell}, 
-	year={2022}, 
-	month={Dec} 
 }
