### 2022-12-30 5d878f2
Added DOI and finished paper.md
