+++ paper/paper.bib
@@ -18,7 +18,8 @@
   number={3},
   pages={637--654},
   year={1973},
-  publisher={The University of Chicago Press}
+  publisher={The University of Chicago Press},
+  DOI={10.1086/260062}
 }
 
  @book{mcdonald_2013, 
@@ -80,4 +81,13 @@
   author={Culp, Christopher L},
   year={2011},
   publisher={John Wiley \& Sons}
+}
+
+@misc{pound_2022, 
+	title={Black Scholes Delta Hedge in Imperfect Markets}, 
+	author={Pound, Mitchell}, 
+	year={2022}, 
+	note={Presented at Fall Student Research Symposium at Utah State University}, 
+	month={Dec} 
+  URL={https://symposium.foragerone.com/2022-usu-fall-student-research-symposium/presentations/50139}
 }
