+++ paper/paper.md
@@ -39,9 +39,9 @@ Bruno is a Julia [@bezanson2017julia] package that allows for comparison of diff
 
 Bruno allows users to compare different financial derivatives hedging and trading strategies. One of the major benefits of Bruno is that it can calculate theoretical historical derivative prices for a variety of pricing models such as the Black-Scholes model[@black1973pricing] and Monte Carlo analysis[@clewlow_strickland_1998]. This is important because derivative price data is not publicly available. Thus, many trading and hedging strategies, by necessity, are currently based on asset prices. Bruno will allow them to be based on theoretical derivative prices instead.
 
-Another key feature of Bruno is that it has the ability to produce a distribution of maximum loss that could result from a trading or hedging strategy. This information would be valuable to financial practitioners and market makers as it would help to quantifying the risk of a potential strategy before putting the strategy into place. Furthermore, it would allow for comparison of different trading and hedging strategies. Creation of these distributions is facilitated by Bruno’s data generating processes. These processes include non-parametric methods, such as the stationary bootstrap [@politis1994stationary] with automatic block-length selection [@politis2004automatic][@patton2009correction] or user defined block lengths, as well as parametric methods such as log diffusion.
+Another key feature of Bruno is that it has the ability to produce a distribution of maximum loss that could result from a trading or hedging strategy. This information would be valuable to financial practitioners and market makers as it would help to quantifying the risk of a potential strategy before putting the strategy into place. Furthermore, it would allow for comparison of different trading and hedging strategies. Creation of these distributions is facilitated by Bruno’s data generating processes. These processes include non-parametric methods, such as the stationary bootstrap [@politis1994stationary] with automatic block-length selection [@politis2004automatic][@patton2009correction] as well as parametric methods such as log diffusion.
 
-Bruno was designed to be used by finance professionals and academics alike. Financial analysis of trading and hedging strategies can be intensive. This package is intended to make this type of investigation more straightforward and accessible. There are many other software packages that have the capacity to calculate derivative prices, simulate hedging, and generate data. However, none of them have been compiled in a manner that allows for complete analysis. Rather, each package performs one part of the process independently, and must be assembled by the software user. Bruno is novel because it allows for complete analysis in a single package. Bruno was recently used in a conference publication, with several other publications nearing completion.
+Bruno was designed to be used by finance professionals and academics alike. Financial analysis of trading and hedging strategies can be intensive. This package is intended to make this type of investigation more straightforward and accessible. There are many other software packages that have the capacity to calculate derivative prices, simulate hedging, and generate data. However, none of them have been compiled in a manner that allows for complete analysis. Rather, each package performs one part of the process independently, and must be assembled by the software user. Bruno is novel because it allows for complete analysis in a single package. Bruno was recently used in a conference publication[@pound_2022], with several other publications nearing completion.
 
 # Acknowledgements
 
