+++ paper/paper.bib
@@ -88,6 +88,6 @@
 	author={Pound, Mitchell}, 
 	year={2022}, 
 	note={Presented at Fall Student Research Symposium at Utah State University}, 
-	month={Dec} 
-  URL={https://symposium.foragerone.com/2022-usu-fall-student-research-symposium/presentations/50139}
+	month={Dec},
+  howpublished={\url{https://symposium.foragerone.com/2022-usu-fall-student-research-symposium/presentations/50139}}
 }
