### 2022-12-30 #5018 @mitchellpound
@jbytecode - Thank you so much! I've linked the citations and the draft paper now generates with the references. It should be ready to go on to next steps.
