### 2022-12-30 #5018 @jbytecode
👋👋👋 Dear @taqtiqa-mark and @lrnv    👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**Bruno: A Julia package for simulation, financial asset pricing and delta hedging**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/5018).

Thank you in advance!
