### 2023-01-01 #5018 @jbytecode
👋👋👋 Dear @daviddewhurst    👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**Bruno: A Julia package for simulation, financial asset pricing and delta hedging**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/5018).

We have already assigned the first reviewer. Hopefully you will be the second one.

Thank you in advance!
