### 2023-01-08 #5018 @omendezmorales
Hi @jbytecode ,

Surely I'm interested, and available for the review. In the past, I've reviewed a paper for a [PyPortfolioOpt ](https://github.com/openjournals/joss-reviews/issues/3066) published also through JOSS, and hence I'm reasonably familiar with the reviewing process.

Please do let me know how  do we proceed.

Kind regards,
