### 2023-01-08 #5056 @editorialbot
[REVIEW]: Bruno: A Julia package for simulation, financial asset pricing and delta hedging
**Submitting author:** @mitchellpound (<a href="http://orcid.org/0000-0002-1244-427X">Mitchell Pound</a>)
**Repository:** https://github.com/USU-Analytics-Solution-Center/Bruno.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.2
