### 2023-01-22 #5056 @jbytecode
Hi dear reviewers @lrnv and @omendezmorales

After 2 weeks since the review starts I just wanted to see how is your review going. As you know, the review process in JOSS in interactive and you always interact with the authors, reviewers, and the editor; you can open issues and send pull requests in the target repo.

Please do not hesitate to ask me anything if you get into trouble.

Thank you in advance!
