### 2023-01-22 #5056 @omendezmorales
Hi Mehmet,

I started checking off some of the items in the rubric, and soon will
contact the author concerning some items I feel he has not addressed (yet).

I will let you know in case of questions/doubts from my side.

Best,

--
/Orlando Méndez/

/
/
