### 2023-01-24 #5056 @omendezmorales
Hi @mitchellpound ,

I'm contacting you in the context of your [submission to JOSS](https://github.com/openjournals/joss-reviews/issues/5056#issue-1524579792) for the paper on _Bruno.jl_. Looking at the review [checklist](https://github.com/openjournals/joss-reviews/issues/5056#issuecomment-1374879708) , there are a few things I feel your paper needs to address/clarify:
* In line 5, you mention "_These authors contributed equally_", meaning Spencer Clemens and you. Is this note important in the paper? and if so, for whom?
* _State of the field_. I think you succeed in describing why Bruno is needed, and what are its key features; however, I think it would be useful to cite some (software) packages Bruno competes with or is superior to (you hint some of this in lines 34-36, but concrete names might help).
* In lines 36 - 38 you mention _complete analysis_ twice; what does _complete analysis_ mean? is it the notion of _simulation, (asset) pricing, and (delta) hedging_, or something else?
* Lastly, I saw the [documentation](https://usu-analytics-solution-center.github.io/Bruno.jl/) is very thorough and easy to folllow, and I believe it would not hurt if you added a (simple) use example in your paper as well (I've seen other JOSS articles doing that, piquing thereby the reader's interest).

I still have some more items to cover in my reviewing role, and I will let you know should I have more questions/comments.

Thanks for your reply!
