### 2023-01-28 ebc4f74
Added example usage and clarifications to paper.md
