+++ paper/paper.md
@@ -8,19 +8,14 @@ tags:
   - Simulation
 authors:
   - name: Mitchell Pound
-    equal-contrib: true
     affiliation: 1
   - name: Spencer Clemens
-    equal-contrib: true 
     affiliation: 1
   - name: Tyler Brough
-    equal-contrib: false
     affiliation: 1
   - name: Pedram Jahangiry
-    equal-contrib: false
     affiliation: 1
   - name: Janette Goodridge
-    equal-contrib: false
     affiliation: 1
 affiliations:
  - name: Utah State University
@@ -41,7 +36,78 @@ Bruno allows users to compare different financial derivatives hedging and tradin
 
 Another key feature of Bruno is that it has the ability to produce a distribution of maximum loss that could result from a trading or hedging strategy. This information would be valuable to financial practitioners and market makers as it would help to quantifying the risk of a potential strategy before putting the strategy into place. Furthermore, it would allow for comparison of different trading and hedging strategies. Creation of these distributions is facilitated by Bruno’s data generating processes. These processes include non-parametric methods, such as the stationary bootstrap [@politis1994stationary] with automatic block-length selection [@politis2004automatic][@patton2009correction] as well as parametric methods such as log diffusion.
 
-Bruno was designed to be used by finance professionals and academics alike. Financial analysis of trading and hedging strategies can be intensive. This package is intended to make this type of investigation more straightforward and accessible. There are many other software packages that have the capacity to calculate derivative prices, simulate hedging, and generate data. However, none of them have been compiled in a manner that allows for complete analysis. Rather, each package performs one part of the process independently, and must be assembled by the software user. Bruno is novel because it allows for complete analysis in a single package. Bruno was recently used in a conference publication[@pound_2022], with several other publications nearing completion.
+Bruno was designed to be used by finance professionals and academics alike. Financial analysis of trading and hedging strategies can be intensive. This package is intended to make this type of investigation more straightforward and accessible. There are many other software packages that have the capacity to calculate derivative prices, simulate hedging, and generate data. For example, in Julia, FinancialDerivatives.jl, FinancialMonteCarlo.jl, and Strategems.jl are packages that can be used for derivative asset pricing, data simulation, and strategy testing, respectively. However, none of them have been compiled in a manner that allows for integrated analysis. Rather, each package performs one part of the process independently, and must be assembled by the software user. Bruno is novel because it allows for derviatives analysis using all these tools in a single package designed to work together. Bruno was recently used in a conference publication[@pound_2022], with several other publications nearing completion.
+
+# Example usage
+
+## Defining a strategy
+Here we demonstrate how to define and use a trading strategy for testing. In this example, a strategy is defined where a derivate asset and it's underlying stock is bought every Friday and held until the end of the month. The `buy` and `sell` functions are provide by Bruno to make defining a strategy easier.
+
+```julia 
+using Bruno
+
+# creating a new strategy subtype for dispatch
+primitive type ExampleStrategy <: Hedging 8 end
+
+
+# define the new strategy 
+import Bruno: buy, sell, strategy
+function Bruno.strategy(fin_obj, 
+                pricing_model, 
+                strategy_mode::Type{ExampleStrategy},
+                holdings,
+                step;
+                kwargs...)
+
+    if step % 5 == 0
+        # buy one FinancialInstrument every 5 days with no transaction costs
+        buy(fin_obj, 1, holdings, pricing_model, 0) 
+        # buy one Stock every 5 days
+        buy(fin_obj.widget, 1, holdings, pricing_model, 0) 
+    end
+
+    return holdings
+end
+```
+
+## Setting up assets and running the strategy
+
[…]
