### 2023-01-28 #5056 @mitchellpound
@omendezmorales thank you for the suggestions!

I made some changes to the paper to clarify the points you brought up:

- I took out the "contributed equally" tag for the authors. We copied that from the JOSS documentation example paper, but after looking at other JOSS papers that were accepted it didn't seem necessary.
- I changed "complete analysis" to be more specific that we meant analysis that used different pricing models and data simulation methods together. Additionally, I referred some financial packages in Julia that have similar functionality of individual parts of Bruno.
- Lastly, I added an example of running a simple trading strategy.

Thanks again for the suggestions, I think each of these changes made the paper much better and showcase the benefits of Bruno.jl much better. If you have any other suggestions please let me know!
