### 2023-02-10 USU-Analytics-Solution-Center/Bruno.jl#61 @lrnv
Documentation code not running
Code from this page does not run : https://usu-analytics-solution-center.github.io/Bruno.jl/tutorials/get_data/get_data/

I thin there is a typing mistake in the `nTimeSteps` -> shoul dnot take an s.

More generally, could you ensure programatically taht code exposed in the documentation at least does not errors when ran ? I think there are pieplines and CI availiable to ensure that.
