### 2023-02-10 USU-Analytics-Solution-Center/Bruno.jl#62 @lrnv
Unclear paper sentence
In the paper, you wrote :

```
There are many other software packages that have the capacity to calculate derivative prices, simulate hedging, and generate data. For example, in Julia, FinancialDerivatives.jl, FinancialMonteCarlo.jl, and Strategems.jl are packages that can be used for derivative asset pricing, data simulation, and strategy testing, respectively. However, none of them have been compiled in a manner that allows for integrated analysis. Rather, each package performs one part of the process independently, and must be assembled by the software user. Bruno is novel because it allows for derviatives analysis using all these tools in a single package designed to work together.
```

"using all these tools" : I thaught at first that "thesse tools" refers to the packages from the previous sentence, and that Bruno was in some sort a Glue package with interfaces from them.

But then, looking at the `Project.toml` I saw that no dependency on these packages is there. Could you please tell me :

- Is bruno allowing me to use all these package together ?
- Or is it replacing them by a fully-integrated tool that does most of what they did separately, but in a unified manner ?

In both case, the sentence from the paper should be fixed IMHO.
