### 2023-02-10 USU-Analytics-Solution-Center/Bruno.jl#63 @lrnv
Short statement of need in the documentation introductory page & readme.
You might include a relatively shorted out version of the JOSS paper, including the statement of need, on this documentation page : https://usu-analytics-solution-center.github.io/Bruno.jl/

Just after the "What is bruno", you may add a "Why Bruno" section that essentially sums up the statement of need from the paper in a sentence or two; that would make the documentation much friendlyer. Also, your readme.md might include this statement too.
