### 2023-02-10 USU-Analytics-Solution-Center/Bruno.jl#64 @lrnv
Merge the readme content and the documentation welcome page content into a unique file. 
It would be nice if the content of the readme (displayed on github) and the welcome page of the documentation are strictly matching each other.

Both of them contain more-or-less the same sections, but with different content.

Would you mind merging them ? I think the best way is to merge them into the readme file and then just include this file in the documentation. Just take the best out of both versions, having a somewhat longer file is not a problem in my honest oppinon
