### 2023-02-10 USU-Analytics-Solution-Center/Bruno.jl#65 @lrnv
About paper citations
You mention other packages in your paper, namely , FinancialDerivatives.jl, FinancialMonteCarlo.jl, and Strategems.jl; but you do not cite them formally.

Please do :)

PS : If they do not provide a DOI, as it looks like it is the case for FinancialDerivatives.jl, you might ask kindly the authors to provide you a propper citation link, i'm sure they would happily do it :)
