### 2023-02-10 USU-Analytics-Solution-Center/Bruno.jl#66 @lrnv
About paper authorship
I see 5 authors to the joss paper and only 4 contributors on github. It is probably fine, but i would like to know the jusitfication for the appearance of the fifth name.

Thanks !
