### 2023-02-10 #5056 @lrnv
Hey,

I have treated the list of items and ended up oppening a few issues on the repo :

- https://github.com/USU-Analytics-Solution-Center/Bruno.jl/issues/61
- https://github.com/USU-Analytics-Solution-Center/Bruno.jl/issues/62
- https://github.com/USU-Analytics-Solution-Center/Bruno.jl/issues/63
- https://github.com/USU-Analytics-Solution-Center/Bruno.jl/issues/64
- https://github.com/USU-Analytics-Solution-Center/Bruno.jl/issues/65
- https://github.com/USU-Analytics-Solution-Center/Bruno.jl/issues/66

A part from these little issue, which are all pretty minor, I found the package easy to use and priced a few calls with it !

I am going to take another shot at using it tomorow and if everything strage pops up i'll say it.

There a still a few boxes that are unchecked, but they should be after those issues are resolved by the authors.

@mitchellpound Bravo, that is a nice little piece of code you have there.
