### 2023-02-11 cc8171a
Typos and clarity fix
