### 2023-02-11 de18b6d
Added package citations
