+++ paper/paper.bib
@@ -1,4 +1,4 @@
- @article{bezanson2017julia,
+@article{bezanson2017julia,
   title={Julia: A fresh approach to numerical computing},
   author={Bezanson, Jeff and Edelman, Alan and Karpinski, Stefan and Shah, Viral B.},
   journal={SIAM Review},
@@ -22,7 +22,7 @@
   DOI={10.1086/260062}
 }
 
- @book{mcdonald_2013, 
+@book{mcdonald_2013, 
  	place={Boston}, 
 	title={Derivatives markets}, 
 	ISBN={9780321543080}, 
@@ -31,7 +31,7 @@
 	year={2013} 
 }
 
- @book{clewlow_strickland_1998, 
+@book{clewlow_strickland_1998, 
  	title={Implementing Derivative Models}, 
 	ISBN={978-0-471-96651-7}, 
 	publisher={John Wiley & Sons Incorporated}, 
@@ -90,4 +90,25 @@
 	note={Presented at Fall Student Research Symposium at Utah State University}, 
 	month={Dec},
   howpublished={\url{https://symposium.foragerone.com/2022-usu-fall-student-research-symposium/presentations/50139}}
+}
+
+@misc{financialderivativesjl,
+  title = {FinancialDerivatives.jl},
+  author = {Brilhante, Alexandre},
+  year = {2018},
+  howpublished = {\url{https://github.com/JuliaQuant/FinancialDerivatives.jl}}
+}
+
+@misc{financialmontecarlojl,
+  title = {FinancialMonteCarlo.jl},
+  author = {Scaramuzzino, Nicola},
+  year = {2018},
+  howpublished = {\url{https://github.com/rcalxrc08/FinancialMonteCarlo.jl}}
+}
+
+@misc{strategemsjl,
+  title = {Strategems.jl},
+  author = {Amos, Jacob},
+  year = {2019},
+  howpublished = {\url{https://github.com/dysonance/Strategems.jl}}
 }
