### 2023-02-11 8a1d56f
changed manual/data_generators.md to have doctests
