### 2023-02-12 c5ccba7
tutorial/fin_inst/derivatives.md to have doctests
