### 2023-02-13 ccfefe9
changed tutorial/strategy/ to have doctests
