+++ paper/paper.bib
@@ -96,19 +96,25 @@
   title = {FinancialDerivatives.jl},
   author = {Brilhante, Alexandre},
   year = {2018},
-  howpublished = {\url{https://github.com/JuliaQuant/FinancialDerivatives.jl}}
+  journal = {Github repository},
+  publisher = {Github},
+  url = {https://github.com/JuliaQuant/FinancialDerivatives.jl}
 }
 
 @misc{financialmontecarlojl,
   title = {FinancialMonteCarlo.jl},
   author = {Scaramuzzino, Nicola},
   year = {2018},
-  howpublished = {\url{https://github.com/rcalxrc08/FinancialMonteCarlo.jl}}
+  journal = {Github repository},
+  publisher = {Github},
+  url = {https://github.com/rcalxrc08/FinancialMonteCarlo.jl}
 }
 
 @misc{strategemsjl,
   title = {Strategems.jl},
   author = {Amos, Jacob},
   year = {2019},
-  howpublished = {\url{https://github.com/dysonance/Strategems.jl}}
+  journal = {Github repository},
+  publisher = {Github},
+  url = {https://github.com/dysonance/Strategems.jl}
 }
