### 2023-02-13 940f238
Made README and docs homepage to be more similar
