+++ README.md
@@ -1,6 +1,11 @@
+# Welcome to Bruno.jl!
+
+## What is Bruno?
 Bruno is a modular, flexible package for simulating financial data, asset pricing, and trading strategy testing. 
 
 Bruno is written and maintained by the [Analytics Solution Center](https://huntsman.usu.edu/asc/index) at Utah State University.  
+
+## Why Bruno? 
 
 ## Contents
 * [Installation instructions](#installation-instructions)
@@ -11,17 +16,18 @@ Bruno is written and maintained by the [Analytics Solution Center](https://hunts
 
 ## Installation Instructions
 
-is a [registered Julia package](https://julialang.org/packages/). So to install it,
-
 1. [Download Julia](https://julialang.org/downloads/).
-
-2. Launch Julia and type
+2. Bruno is a [registered Julia package](https://julialang.org/packages/). So to install it, Launch Julia and type:
 
 ```julia
 julia> using Pkg
 
 julia> Pkg.add("Bruno")
 ```
+
+Note: The latest version of Bruno strongly suggests _at least_ Julia 1.8 or later to run.
+While most scripts will run on Julia 1.6 or 1.7, Bruno is _only_ tested on Julia 1.8.
+
 ## What Bruno can do
 Bruno's four main functions are 
 * Describe financial tools like stock options and commodity futures with an intuitive type system
@@ -37,7 +43,15 @@ call = EuroCallOption(stock, 55; maturity=.5)
 price!(call, BlackScholes)
 ```
 
-Check out [the documentation](https://usu-analytics-solution-center.github.io/Bruno.jl/) for more examples and tutorials. 
+Check out [the tutorials](https://usu-analytics-solution-center.github.io/Bruno.jl/) for more examples. 
+
+## Getting In Touch
+
+Whether you need help getting started with Bruno, found a bug, want Bruno to be even better, or just want to chat about computational economics and finance, there are two ways of contacting us:
+
+* [Start a discussion](https://github.com/USU-Analytics-Solution-Center/Bruno.jl/discussions). This is great for general questions about numerics, computation finance, experimental or under-documented features, and for getting help setting up a cool new trading strategy simulation.
+* [Open an issue](https://github.com/USU-Analytics-Solution-Center/Bruno.jl/issues). Issues are best if you think the Bruno source code needs attention: such as a bug, a type inconsistency error (😱), an important missing feature, or a typo in this documentation 👀.
+
 
 ## Citing
 If you use Bruno.jl as part of your research, teaching, or other activities, we would be grateful if you could cite our work and mention Bruno.jl by name.
