### 2023-02-13 11e1346
Added Why Bruno section to readme and docs homepage
