+++ README.md
@@ -5,7 +5,15 @@ Bruno is a modular, flexible package for simulating financial data, asset pricin
 
 Bruno is written and maintained by the [Analytics Solution Center](https://huntsman.usu.edu/asc/index) at Utah State University.  
 
-## Why Bruno? 
+## What Bruno can do
+Bruno's four main functions are 
+* Describe financial tools like stock options and commodity futures with an intuitive type system
+* Generate simulated time-series data with parametric and non-parametric models
+* Price complex assets
+* Test trading or hedging strategies with historical or simulated prices
+
+## Why Bruno?
+Financial analysis of trading and hedging strategies can be intensive. This package is intended to make this type of investigation more straightforward and accessible. There are many other software packages that have the capacity to calculate derivative prices, simulate hedging, or generate data. However, none of them have been compiled in a manner that allows for integrated analysis. Rather, each package performs one part of the process independently, and must be assembled by the software user. In contrast, Bruno is novel because it provides a replacement to these independent packages with a fully integrated set of tools for derivatives analysis designed to work in a unified manner in Julia.
 
 ## Contents
 * [Installation instructions](#installation-instructions)
@@ -27,14 +35,6 @@ julia> Pkg.add("Bruno")
 
 Note: The latest version of Bruno strongly suggests _at least_ Julia 1.8 or later to run.
 While most scripts will run on Julia 1.6 or 1.7, Bruno is _only_ tested on Julia 1.8.
-
-## What Bruno can do
-Bruno's four main functions are 
-* Describe financial tools like stock options and commodity futures with an intuitive type system
-* Generate simulated time-series data with parametric and non-parametric models
-* Price complex assets
-* Test trading or hedging strategies with historical or simulated prices
-
 ## Pricing your first asset
 ```julia
 using Bruno
@@ -43,7 +43,7 @@ call = EuroCallOption(stock, 55; maturity=.5)
 price!(call, BlackScholes)
 ```
 
-Check out [the tutorials](https://usu-analytics-solution-center.github.io/Bruno.jl/) for more examples. 
+Check out [the tutorials](https://usu-analytics-solution-center.github.io/Bruno.jl/tutorials/fin_inst/base_asset/) for more examples. 
 
 ## Getting In Touch
 
@@ -51,7 +51,6 @@ Whether you need help getting started with Bruno, found a bug, want Bruno to be 
 
 * [Start a discussion](https://github.com/USU-Analytics-Solution-Center/Bruno.jl/discussions). This is great for general questions about numerics, computation finance, experimental or under-documented features, and for getting help setting up a cool new trading strategy simulation.
 * [Open an issue](https://github.com/USU-Analytics-Solution-Center/Bruno.jl/issues). Issues are best if you think the Bruno source code needs attention: such as a bug, a type inconsistency error (😱), an important missing feature, or a typo in this documentation 👀.
-
 
 ## Citing
 If you use Bruno.jl as part of your research, teaching, or other activities, we would be grateful if you could cite our work and mention Bruno.jl by name.
@@ -71,4 +70,4 @@ We are always looking to expand the community and make Bruno better.
 
 If you'd like to work on a new feature, or if you're new to open source and want to crowd-source neat projects that fit your interests, you should [start a discussion](https://github.com/USU-Analytics-Solution-Center/Bruno.jl//discussions/new?) right away.
 
-For more information check out our contributor's guide.
+For more information check out our [contributor's guide](https://github.com/USU-Analytics-Solution-Center/Bruno.jl/blob/main/CONTRIBUTING.md).
