### 2023-02-13 6b4ea53
Removed contents section from README
