### 2023-02-14 68314ae
Changed all structs to not use abstract types
