+++ src/DataGeneration/bootstrap.jl
@@ -15,8 +15,8 @@ function. T can be any subtype of TSBootMethod: Stationary, MovingBlock, or Circ
 ## Keyword Arguments
 - `input_data::Array{<:Real}`: data to be resampled. Must be a 1-D array
 - `bootstrap_method`: Type of time series bootstrap to use. Must be subtype of TSBootMethod.
-- `n::Integer`: size of resampled output data. Default: 100
-- `block_size::Integer`: block size to use. Defaults to the optimal block length using `opt_block_length()`
+- `n::Int64`: size of resampled output data. Default: 100
+- `block_size::Float64`: block size to use. Defaults to the optimal block length using `opt_block_length()`
 
 ## Examples
 ```julia
@@ -30,8 +30,8 @@ input2 = BootstrapInput{MovingBlock}(;kwargs...)
 """
 struct BootstrapInput{T<:TSBootMethod} <: DataGenInput
     input_data::Array{<:Real} # array of data to be resampled
-    n::Integer # desired size of resampled data
-    block_size::Float32 #desired average block size (will add more to this later)
+    n::Int64 # desired size of resampled data
+    block_size::Float64 #desired average block size (will add more to this later)
 
     # constructor for kwargs
     function BootstrapInput{T}(;
