+++ src/Instruments/widgets.jl
@@ -14,23 +14,23 @@ abstract type Widget end
 
 Widget subtype. Used as a base or root asset for financial instrument.
 """
-struct Stock <: Widget
-    prices::Array{AbstractFloat}
+struct Stock{T} <: Widget
+    prices::Vector{T}
     name::String
-    timesteps_per_period::Integer
-    volatility::AbstractFloat
+    timesteps_per_period::Int64
+    volatility::Float64
 
     # constructor for kwargs
 
-    function Stock(;
+    function Stock{T}(;
         prices,
         name = "",
         timesteps_per_period = length(prices),
         volatility = get_volatility(prices, timesteps_per_period),
         _...
-    )
+    ) where T
         # allows single price input through kwargs (and ordered arguments)
-        if typeof(prices) <: Real
+        if typeof(prices) <: Number
             prices >= 0 ? prices = [prices] :
             error("Single price point must be non-negative")
             volatility == nothing ?
@@ -45,16 +45,16 @@ struct Stock <: Widget
         # catch negative timesteps_per_period
         timesteps_per_period >= 0 ? nothing : 
         error("timesteps_per_period cannot be negative")
-        new(prices, name, timesteps_per_period, volatility)
+	new{T}(prices, name, timesteps_per_period, volatility)
     end
 
     # constructor for ordered argumentes 
-    function Stock(
+    function Stock{T}(
         prices,
         name = "",
         timesteps_per_period = length(prices),
         volatility = get_volatility(prices, timesteps_per_period)
-    )
+    ) where T
         if typeof(prices) <: Real
             prices >= 0 ? prices = [prices] :
             error("Single price point must be non-negative")
@@ -70,7 +70,7 @@ struct Stock <: Widget
         timesteps_per_period >= 0 ? nothing : 
         # catch negative timesteps_per_period
         error("timesteps_per_period cannot be negative")
-        new(prices, name, timesteps_per_period, volatility)
+	new{T}(prices, name, timesteps_per_period, volatility)
     end
 end
 
@@ -85,7 +85,7 @@ Construct a Stock type to use as a base asset for FinancialInstrument.
 ## Arguments
 - `prices`:Historical prices (input as a 1-D array) or the current price input as a number `<: Real`
 - `name::String`: Name of the stock or stock ticker symbol. Default "".
-- `timesteps_per_period::Integer`: For the size of a timestep in the data, the number of 
+- `timesteps_per_period::Int64`: For the size of a timestep in the data, the number of 
 time steps for a given period of time, cannot be negative. For example, if the period of 
 interest is a year, and daily stock data is used, `timesteps_per_period=252`. Defualt is 
 length of the `prices` array or 0 for single price (static) stock. 
@@ -113,8 +113,22 @@ Stock(40; volatility=.05)
 """
 function Stock(price::Real; name = "", volatility)
     prices = [price]
-    Stock(; prices = prices, name = name, volatility = volatility, timesteps_per_period = 0)
+    Stock{typeof(price)}(; prices = prices, name = name, volatility = volatility, timesteps_per_period = 0)
 end
+# outer constructor to infer the type used in the prices array
+Stock(
+    prices,
+    name = "",
+    timesteps_per_period = length(prices),
+    volatility = get_volatility(prices, timesteps_per_period)
+) = Stock{eltype(prices)}(prices, name, timesteps_per_period, volatility)
+Stock(;
+        prices,
+        name = "",
+        timesteps_per_period = length(prices),
+        volatility = get_volatility(prices, timesteps_per_period),
[…]
