+++ src/Instruments/financial_instruments.jl
@@ -34,12 +34,11 @@ European call option with underlying asset `T`.
 """
 struct EuroCallOption{T<:Widget} <: CallOption{T}
     widget::T
-    strike_price::AbstractFloat
-    maturity::AbstractFloat
-    risk_free_rate::AbstractFloat
-
+    strike_price::Float64
+    maturity::Float64
+    risk_free_rate::Float64
     label::String
-    values_library::Dict{String,Dict{String,AbstractFloat}}
+    values_library::Dict{String,Dict{String,Float64}}
 
     # kwargs constructor
     function EuroCallOption{T}(;
@@ -48,11 +47,11 @@ struct EuroCallOption{T<:Widget} <: CallOption{T}
         maturity = 1,
         risk_free_rate = 0.02,
         label = "",
-        values_library = Dict{String,Dict{String,AbstractFloat}}(),
+        values_library = Dict{String,Dict{String,Float64}}(),
     ) where {T<:Widget}
         strike_price >= 0 ? nothing : error("strike_price must be non-negative")
         maturity >= 0 ? nothing : error("maturity must be positive ", maturity)
-        values_library == Dict{String,Dict{String,AbstractFloat}}() ? nothing :
+        values_library == Dict{String,Dict{String,Float64}}() ? nothing :
         @warn("It is not recommended to pass values through the constructor. \
         price!(Instrument, pricing_model) should be used")
         new{T}(widget, strike_price, maturity, risk_free_rate, label, values_library)
@@ -69,7 +68,7 @@ struct EuroCallOption{T<:Widget} <: CallOption{T}
     ) where {T<:Widget}
         strike_price >= 0 ? nothing : error("strike_price must be non-negative")
         maturity >= 0 ? nothing : error("maturity must be positive")
-        values_library == Dict{String,Dict{String,AbstractFloat}}() ? nothing :
+        values_library == Dict{String,Dict{String,Float64}}() ? nothing :
         @warn("It is not recommended to pass values through the constructor. \
         price!(Instrument, pricing_model) should be used")
 
@@ -110,7 +109,7 @@ EuroCallOption(
     maturity = 1,
     risk_free_rate = 0.02,
     label = "",
-    values_library = Dict{String,Dict{String,AbstractFloat}}(),
+    values_library = Dict{String,Dict{String,Float64}}(),
 ) = EuroCallOption{typeof(widget)}(;
     widget = widget,
     strike_price = strike_price,
@@ -125,7 +124,7 @@ EuroCallOption(
     strike_price::Real,
     maturity::Real,
     label::String,
-    values_library::Dict{String,Dict{String,AbstractFloat}},
+    values_library::Dict{String,Dict{String,Float64}},
 ) = EuroCallOption{typeof(widget)}(;
     widget = widget,
     strik_price = strike_price,
@@ -140,7 +139,7 @@ EuroCallOption(;
     maturity = 1,
     risk_free_rate = 0.02,
     label = "",
-    values_library = Dict{String,Dict{String,AbstractFloat}}(),
+    values_library = Dict{String,Dict{String,Float64}}(),
 ) = EuroCallOption{typeof(widget)}(;
     widget = widget,
     strike_price = strike_price,
@@ -157,12 +156,12 @@ American call option with underlying asset `T`.
 """
 struct AmericanCallOption{T<:Widget} <: CallOption{T}
     widget::T
-    strike_price::AbstractFloat
-    maturity::AbstractFloat
-    risk_free_rate::AbstractFloat
+    strike_price::Float64
+    maturity::Float64
+    risk_free_rate::Float64
 
     label::String
-    values_library::Dict{String,Dict{String,AbstractFloat}}
[…]

[…]
