+++ search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"tutorials/fin_inst/derivatives/#Derivatives-(and-Other-Financial-Instruments)","page":"Derivatives (and Other Financial Instruments)","title":"Derivatives (and Other Financial Instruments)","text":"","category":"section"},{"location":"tutorials/fin_inst/derivatives/#Overview","page":"Derivatives (and Other Financial Instruments)","title":"Overview","text":"","category":"section"},{"location":"tutorials/fin_inst/derivatives/","page":"Derivatives (and Other Financial Instruments)","title":"Derivatives (and Other Financial Instruments)","text":"Financial instruments are financial tools that depend on an underlying asset (or eventually more than one). Common examples are call and put options on stocks and commodities. They can be used in combination with both the price! and the strategy_returns functions.","category":"page"},{"location":"tutorials/fin_inst/derivatives/#Creating-a-stock-option","page":"Derivatives (and Other Financial Instruments)","title":"Creating a stock option","text":"","category":"section"},{"location":"tutorials/fin_inst/derivatives/","page":"Derivatives (and Other Financial Instruments)","title":"Derivatives (and Other Financial Instruments)","text":"Here's how to create a stock Option (more specifically a European Call Option). To do that, we need to make a stock first. You can find a tutorial here.","category":"page"},{"location":"tutorials/fin_inst/derivatives/","page":"Derivatives (and Other Financial Instruments)","title":"Derivatives (and Other Financial Instruments)","text":"historical_prices = [1, 2, 3, 4, 5]\n\n# creates a Stock widget assuming hisorical prices are daily prices\na_widget = Stock(;prices=historical_prices, name=\"my_widget\", timesteps_per_period=252)\n\n# creates a call option with a strike price of $60\ncall = EuroCallOption(a_widget, 60)","category":"page"},{"location":"tutorials/fin_inst/derivatives/#Interacting-With-a-Financial-Instrument-(Asset-Pricing)","page":"Derivatives (and Other Financial Instruments)","title":"Interacting With a Financial Instrument (Asset Pricing)","text":"","category":"section"},{"location":"tutorials/fin_inst/derivatives/","page":"Derivatives (and Other Financial Instruments)","title":"Derivatives (and Other Financial Instruments)","text":"One of the more interesting things that can be done with financial instruments is asset pricing with several models. Not all of the models work with all the financial instruments, so check the manual to see if it is supported and a method is written for each model.","category":"page"},{"location":"tutorials/fin_inst/derivatives/","page":"Derivatives (and Other Financial Instruments)","title":"Derivatives (and Other Financial Instruments)","text":"# pricing the EuroCallOption from above\n# values will be written to values_library in the option\nprice!(call, BlackScholes)\n\n# accessing the value from values_library\ncall.values_library[\"BlackScholes\"][\"value\"]","category":"page"},{"location":"reference/hedge_simulate/#Hedging-Simulations","page":"Hedging Simulations","title":"Hedging Simulations","text":"","category":"section"},{"location":"reference/hedge_simulate/#trading_strategies","page":"Hedging Simulations","title":"Backtesting and Trading/Hedging simulation","text":"","category":"section"},{"location":"reference/hedge_simulate/","page":"Hedging Simulations","title":"Hedging Simulations","text":"Bruno.buy\nBruno.sell\nstrategy\nstrategy_returns","category":"page"},{"locatio
[…]
