### 2023-02-14 6005275
Took out unnecessary type annotations
