+++ src/Instruments/widgets.jl
@@ -111,7 +111,7 @@ Stock(;kwargs...)
 Stock(40; volatility=.05)
 ```
 """
-function Stock(price::Real; name = "", volatility)
+function Stock(price::Number; name = "", volatility)
     prices = [price]
     Stock{typeof(price)}(; prices = prices, name = name, volatility = volatility, timesteps_per_period = 0)
 end
@@ -234,7 +234,7 @@ Commodity(;kwargs...)
 Commodity(40; volatility=.05)
 ```
 """
-function Commodity(price::Real; name = "", volatility)
+function Commodity(price::Number; name = "", volatility)
     prices = [price]
     Commodity{typeof(price)}(;
         prices = prices,
@@ -309,7 +309,7 @@ Bond(;kwargs...)
 Bond(2; coupon_rate=.05)
 ```
 """
-function Bond(price::Real; name = "", time_mat = 1, coupon_rate = 0.03)
+function Bond(price::Number; name = "", time_mat = 1, coupon_rate = 0.03)
     prices = [price]
     Bond{typeof(price)}(; prices = prices, name = name, time_mat = time_mat, coupon_rate = coupon_rate)
 end
@@ -338,7 +338,7 @@ end
 
 get_volatility(prices) = get_volatility(prices, length(prices))
 
-function add_price_value(a_widget::Widget, a_new_price::Real)
+function add_price_value(a_widget::Widget, a_new_price)
     a_new_price >= 0 ? nothing :
     @warn("You are trying to add a negative number to a prices list")
     push!(a_widget.prices, a_new_price)
