### 2023-02-14 USU-Analytics-Solution-Center/Bruno.jl#66 @mitchellpound
The authors that aren't contributors on GitHub were invaluable in helping to design the architecture of the package and type structure as well as help on the financial theory side of the project. Though they didn't push any code to the repo directly, they often were in pair coding sessions and overall were a large reason the package works :)
