### 2023-02-14 8f6662f
Fixed BootstrapInput parametric struct
