### 2023-02-15 5d77bee
Fixed LogDiffInput struct
