### 2023-02-15 3b9dad8
Fixed Widget structs
