+++ src/Instruments/widgets.jl
@@ -14,23 +14,23 @@ abstract type Widget end
 
 Widget subtype. Used as a base or root asset for financial instrument.
 """
-struct Stock{T} <: Widget
+struct Stock{T, TI, TF} <: Widget
     prices::Vector{T}
     name::String
-    timesteps_per_period::Int64
-    volatility::Float64
+    timesteps_per_period::TI
+    volatility::TF
 
     # constructor for kwargs
 
-    function Stock{T}(;
+    function Stock{T,TI,TF}(;
         prices,
         name = "",
         timesteps_per_period = length(prices),
         volatility = get_volatility(prices, timesteps_per_period),
         _...
-    ) where T
+    ) where {T,TI,TF}
         # allows single price input through kwargs (and ordered arguments)
-        if typeof(prices) <: Number
+        if !isa(prices, Vector)
             prices >= 0 ? prices = [prices] :
             error("Single price point must be non-negative")
             volatility == nothing ?
@@ -45,17 +45,17 @@ struct Stock{T} <: Widget
         # catch negative timesteps_per_period
         timesteps_per_period >= 0 ? nothing : 
         error("timesteps_per_period cannot be negative")
-	new{T}(prices, name, timesteps_per_period, volatility)
+	new{T,TI,TF}(prices, name, timesteps_per_period, volatility)
     end
 
     # constructor for ordered argumentes 
-    function Stock{T}(
+    function Stock{T,TI,TF}(
         prices,
         name = "",
         timesteps_per_period = length(prices),
         volatility = get_volatility(prices, timesteps_per_period)
-    ) where T
-        if typeof(prices) <: Real
+    ) where {T,TI,TF}
+        if !isa(prices, Vector)
             prices >= 0 ? prices = [prices] :
             error("Single price point must be non-negative")
             volatility == nothing ?
@@ -70,7 +70,7 @@ struct Stock{T} <: Widget
         timesteps_per_period >= 0 ? nothing : 
         # catch negative timesteps_per_period
         error("timesteps_per_period cannot be negative")
-	new{T}(prices, name, timesteps_per_period, volatility)
+	new{T,TI,TF}(prices, name, timesteps_per_period, volatility)
     end
 end
 
@@ -111,47 +111,60 @@ Stock(;kwargs...)
 Stock(40; volatility=.05)
 ```
 """
-function Stock(price::Real; name = "", volatility)
+function Stock(price; name = "", volatility)
     prices = [price]
-    Stock{typeof(price)}(; prices = prices, name = name, volatility = volatility, timesteps_per_period = 0)
+    T = typeof(price)
+    TI = Int16
+    return Stock{T,TI,T}(; prices = prices, name = name, volatility = volatility, timesteps_per_period = 0)
 end
+
 # outer constructor to infer the type used in the prices array
-Stock(
-    prices,
+function Stock(
+    prices::Vector,
     name = "",
     timesteps_per_period = length(prices),
     volatility = get_volatility(prices, timesteps_per_period)
-) = Stock{eltype(prices)}(prices, name, timesteps_per_period, volatility)
-Stock(;
+)
+    T = eltype(prices)
+    TI = typeof(timesteps_per_period)
+    TF = typeof(volatility)
+    return Stock{T,TI,TF}(prices, name, timesteps_per_period, volatility)
+end
+function Stock(;
         prices,
         name = "",
         timesteps_per_period = length(prices),
         volatility = get_volatility(prices, timesteps_per_period),
         _...
-) = Stock{eltype(prices)}(prices, name, timesteps_per_period, volatility)
-
+)
[…]
