### 2023-02-16 318cd91
Fixed FinancialInstrument structs
