+++ src/Instruments/financial_instruments.jl
@@ -32,53 +32,51 @@ abstract type PutOption{T<:Widget} <: Option end
 
 European call option with underlying asset `T`. 
 """
-struct EuroCallOption{T<:Widget} <: CallOption{T}
+struct EuroCallOption{T<:Widget,S,D} <: CallOption{T}
     widget::T
-    strike_price::Float64
-    maturity::Float64
-    risk_free_rate::Float64
+    strike_price::S
+    maturity::S
+    risk_free_rate::S
     label::String
-    values_library::Dict{String,Dict{String,Float64}}
+    values_library::Dict{String,Dict{String,D}}
 
     # kwargs constructor
-    function EuroCallOption{T}(;
+    function EuroCallOption{T,S,D}(;
         widget,
         strike_price = widget.prices[end],
         maturity = 1,
         risk_free_rate = 0.02,
         label = "",
         values_library = Dict{String,Dict{String,Float64}}(),
-    ) where {T<:Widget}
+    ) where {T<:Widget,S,D}
         strike_price >= 0 ? nothing : error("strike_price must be non-negative")
         maturity >= 0 ? nothing : error("maturity must be positive ", maturity)
-        values_library == Dict{String,Dict{String,Float64}}() ? nothing :
+        values_library == Dict{String,Dict{String,D}}() ? nothing :
         @warn("It is not recommended to pass values through the constructor. \
         price!(Instrument, pricing_model) should be used")
-        new{T}(widget, strike_price, maturity, risk_free_rate, label, values_library)
+        new{T,S,D}(widget, strike_price, maturity, risk_free_rate, label, values_library)
     end
 
     # ordered arguments constructor
-    function EuroCallOption{T}(
-        widget::T,
+    function EuroCallOption{T,S,D}(
+        widget,
         strike_price,
         maturity,
         risk_free_rate,
         label,
         values_library,
-    ) where {T<:Widget}
+    ) where {T<:Widget,S,D}
         strike_price >= 0 ? nothing : error("strike_price must be non-negative")
         maturity >= 0 ? nothing : error("maturity must be positive")
-        values_library == Dict{String,Dict{String,Float64}}() ? nothing :
+        values_library == Dict{String,Dict{String,D}}() ? nothing :
         @warn("It is not recommended to pass values through the constructor. \
         price!(Instrument, pricing_model) should be used")
-
-        new{T}(widget, strike_price, maturity, risk_free_rate, label, values_library)
+        new{T,S,D}(widget, strike_price, maturity, risk_free_rate, label, values_library)
     end
 end
 
 # Outer constructors for passing only the widget
 """
-    EuroCallOption(widget, strike_price; kwargs...)
     EuroCallOption(;kwargs...)
     EuroCallOption{T<:Widget}(;kwargs...)
     EuroCallOption{T<:Widget}(widget, strike_price, maturity, risk_free_rate, values_library)
@@ -86,7 +84,7 @@ end
 Construct a EuroCallOption with underlying asset `T`.
 
 ## Arguments
-- `widget::Widget`: underlying asset
+- `widget`: underlying asset
 - `strike_price`: Contracted price to buy underlying asset at maturity
 - `maturity`: time to maturity of the option with respect to implicit time period. Default 1.
 - `risk_free_rate`: market risk free interest rate. Default is .02.
@@ -103,98 +101,83 @@ kwargs = Dict(:widget=>stock, :strike_price=>10, :maturity=>1, :risk_free_rate=>
 EuroCallOption(;kwargs...)
 ```
 """
-EuroCallOption(
-    widget::Widget,
-    strike_price::Real = widget.prices[end];
-    maturity = 1,
[…]
