+++ test/hedgingtests.jl
@@ -6,7 +6,7 @@ using Logging
 @testset "buy function tests" begin
     @testset "buy without transaction costs" begin
         test_stock = Stock(41.0; name = "test", volatility = 0.3)  
-        test_call = EuroCallOption(test_stock, 40; label = "test_call", risk_free_rate = 0.08, maturity = 0.25)
+        test_call = EuroCallOption(;widget = test_stock, strike_price = 40, label = "test_call", risk_free_rate = 0.08, maturity = 0.25)
         # the price!() of this call should be 3.399 (see pricingmodel tests)
         holdings = Dict("cash" => 10.0, "test_call" => 1, "test" => 1)
         # buy call without transaction costs
@@ -26,7 +26,7 @@ using Logging
 
     @testset "buy with transaction_costs" begin
         test_stock = Stock(41.0; name = "test", volatility = 0.3)  
-        test_call = EuroCallOption(test_stock, 40; label = "test_call", risk_free_rate = 0.08, maturity = 0.25)
+        test_call = EuroCallOption(;widget = test_stock, strike_price = 40, label = "test_call", risk_free_rate = 0.08, maturity = 0.25)
         # the price!() of this call should be 3.399 (see pricingmodel tests)
         holdings = Dict("cash" => 10.0, "test_call" => 1, "test" => 1)
         # buy call without transaction costs
@@ -46,7 +46,7 @@ using Logging
 
     @testset "buy function limitations" begin
         test_stock = Stock(41.0; name = "test", volatility = 0.3)  
-        test_call = EuroCallOption(test_stock, 40; label = "test_call", risk_free_rate = 0.08, maturity = 0.25)
+        test_call = EuroCallOption(;widget = test_stock, strike_price = 40, label = "test_call", risk_free_rate = 0.08, maturity = 0.25)
         holdings = Dict("cash" => 10.0, "test_call" => 1, "test" => 1)
         # buying negative widget
         @test_logs (:warn,"unable to buy negative amounts. Use sell instead") buy(test_stock, -1, holdings, BlackScholes)
@@ -56,7 +56,7 @@ using Logging
         @test_logs (:warn,"unable to buy negative amounts. Use sell instead") buy(test_call, -1, holdings, BlackScholes, 1)
 
         # buying expired FinancialInstrument
-        expired_call = EuroCallOption(test_stock, 40; label = "test_call", risk_free_rate = 0.08, maturity = 0.0)
+        expired_call = EuroCallOption(;widget = test_stock, strike_price = 40, label = "test_call", risk_free_rate = 0.08, maturity = 0.0)
         @test_logs (:warn, "unable to buy expired FinancialInstrument") buy(expired_call, 1, holdings, BlackScholes, 1)
     end
 end
@@ -64,7 +64,7 @@ end
 @testset "sell function tests" begin
     @testset "sell without transaction costs" begin
         test_stock = Stock(41.0; name = "test", volatility = 0.3)  
-        test_call = EuroCallOption(test_stock, 40; label = "test_call", risk_free_rate = 0.08, maturity = 0.25)
+        test_call = EuroCallOption(;widget = test_stock, strike_price = 40, label = "test_call", risk_free_rate = 0.08, maturity = 0.25)
         # the price!() of this call should be 3.399 (see pricingmodel tests)
         holdings = Dict("cash" => 10.0, "test_call" => 1, "test" => 1)
         # sell call without transaction costs
@@ -84,7 +84,7 @@ end
 
     @testset "sell with transaction_costs" begin
[…]
