### 2023-02-16 938779b
Removed useless type annotations from methods
