+++ src/DataGeneration/factory.jl
@@ -19,7 +19,7 @@ widget = Stock(prices)
 list_of_widgets = factory(widget, Stationary, 2)
 ```
 """
-function factory(widget::Widget, bootstrap_method::Type{<:TSBootMethod}, nWidgets::Signed)
+function factory(widget::Widget, bootstrap_method, nWidgets)
     fields = field_exclude(widget)
     # calculating the returns
     # TODO: check if time series Stationary 
@@ -52,8 +52,8 @@ end
 
 function factory(
     fin_instrument::FinancialInstrument,
-    bootstrap_method::Type{<:TSBootMethod},
-    nInstruments::Signed,
+    bootstrap_method,
+    nInstruments,
 )
     fields = field_exclude(fin_instrument)
     widget_ar = factory(fin_instrument.widget, bootstrap_method, nInstruments)
