### 2023-02-16 2f14394
Fixed other type annotations that I missed
