### 2023-02-16 10a40fd
Changed some conversions to promotions
