### 2023-02-16 df8366b
Fixed type instability in pricing and widgets
