+++ src/BackTest/hedging.jl
@@ -74,13 +74,13 @@ function strategy_returns(
     pricing_model,
     strategy_type,
     future_prices,
-    n_timesteps::Int,
-    timesteps_per_period::Int,
-    cash_injection::AbstractFloat = 0,
-    fin_obj_count::AbstractFloat = 0,
-    widget_count::AbstractFloat = 0,
-    pay_int_rate::AbstractFloat = 0,
-    hold_return_int_rate::AbstractFloat = 0;
+    n_timesteps,
+    timesteps_per_period,
+    cash_injection = 0.0,
+    fin_obj_count = 0.0,
+    widget_count = 0.0,
+    pay_int_rate = 0.0,
+    hold_return_int_rate = 0.0;
     kwargs...
 ) 
     # make some checks
@@ -156,44 +156,6 @@ function strategy_returns(
     end
 
     return holdings["cash"] - initial_value, ts_holdings, obj
-end
-
-# converter function to let users input whatever type of number they want
-function strategy_returns(
-    obj::FinancialInstrument,
-    pricing_model,
-    strategy_type,
-    future_prices,
-    n_timesteps::Int,
-    timesteps_per_period::Int,
-    cash_injection::Real = 0,
-    fin_obj_count::Real = 0,
-    widget_count::Real = 0,
-    pay_int_rate::Real = 0,
-    hold_return_int_rate::Real = 0;
-    kwargs...
-)  
-
-    cash_injection = convert(AbstractFloat, cash_injection)
-    fin_obj_count = convert(AbstractFloat,fin_obj_count)
-    widget_count= convert(AbstractFloat, widget_count)
-    pay_int_rate = convert(AbstractFloat, pay_int_rate)
-    hold_return_int_rate= convert(AbstractFloat, hold_return_int_rate)
-
-    strategy_returns(
-        obj,
-        pricing_model,
-        strategy_type,
-        future_prices,
-        n_timesteps,
-        timesteps_per_period,
-        cash_injection,
-        fin_obj_count,
-        widget_count,
-        pay_int_rate,
-        hold_return_int_rate;
-        kwargs...
-    )  
 end
 
 # for an array of fin objs
@@ -281,16 +243,16 @@ function strategy_returns(
     objs::Vector{<:FinancialInstrument},
     pricing_model,
     strategy_type,
-    future_prices::Dict{String,Vector{AbstractFloat}},
-    n_timesteps::Int,
-    timesteps_per_period::Int,
-    cash_injection::AbstractFloat = 0.0,
-    fin_obj_count = Dict{String,AbstractFloat}(),
-    widget_count = Dict{String,AbstractFloat}(),
-    pay_int_rate::AbstractFloat = 0,
-    hold_return_int_rate::AbstractFloat = 0;
+    future_prices,
+    n_timesteps,
+    timesteps_per_period,
+    cash_injection = 0.0,
+    fin_obj_count = Dict{String,Float64}(),
+    widget_count = Dict{String,Float64}(),
+    pay_int_rate = 0.0,
+    hold_return_int_rate = 0.0;
     kwargs...
-) where {T<:Real}
+)
 
     # checks before the sim starts
     for fin_obj in objs
@@ -332,12 +294,12 @@ function strategy_returns(
 
 
     # set up holdings dictionary. Holdings is the active holdings of the program while ts_holdings produces a history
-   holdings = Dict{String, AbstractFloat}(
+   holdings = Dict(
         "cash" => cash_injection, 
         fin_obj_count..., 
         widget_count...
     ) 
-    ts_holdings = Dict("cash" => AbstractFloat[cash_injection])
+    ts_holdings = Dict("cash" => [cash_injection])
 
     for key in keys(widget_dict)
         try
[…]
