### 2023-02-16 dd8c598
Changed docstrings to show the simplified methods. Removed type annotations
