### 2023-02-16 00be29d
Took out the last few abstract objects I missed
