+++ test/fininsttests.jl
@@ -21,7 +21,7 @@
         @test test_fininst.maturity == 0.5
         @test test_fininst.risk_free_rate == 0.08
         @test test_fininst.label == "test"
-        @test test_fininst.values_library == Dict{String,Dict{String,AbstractFloat}}()
+        @test test_fininst.values_library == Dict{String,Dict{String,Float64}}()
     end
 
     @testset "Only widget and strike_price constructor for $fininst" for fininst in [
@@ -36,7 +36,7 @@
         @test test_fininst.maturity == 1.0
         @test test_fininst.risk_free_rate == 0.02
         @test test_fininst.label == ""
-        @test test_fininst.values_library == Dict{String,Dict{String,AbstractFloat}}()
+        @test test_fininst.values_library == Dict{String,Dict{String,Float64}}()
     end
     
     @testset "Constructor limits for $fininst" for fininst in [
@@ -56,7 +56,7 @@
         @test_warn "It is not recommended to pass values through the constructor. price!(Instrument, pricing_model) should be used" fininst(;
             widget = widget, 
             strike_price = 6, 
-            values_library=Dict{String, Dict{String, AbstractFloat}}("test" => Dict("test2" => 1.0)))
+            values_library=Dict{String, Dict{String, Float64}}("test" => Dict("test2" => 1.0)))
     end
 end
 
