### 2023-02-16 03dc8ba
The rest of the useless type annotations (after searching entire package) this should be all of them!
