### 2023-02-16 0645bb1
To make tests passing again. Added necessary type annotation for constructor dispatch
