### 2023-02-17 aba2bb0
Final typos and documentation cleanup
