+++ docs/src/tutorials/strategy/trading_strategy.md
@@ -89,7 +89,8 @@ cumulative_returns, holdings = strategy_returns(
 
 # output
 
-(0.3370435884062317, Dict("example_stock" => [0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 1.0, 1.0, 1.0, 1.0  …  2.0, 2.0, 2.0, 3.0, 3.0, 3.0, 3.0, 3.0, 4.0, 0.0], "" => [0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 1.0, 1.0, 1.0, 1.0  …  2.0, 2.0, 2.0, 3.0, 3.0, 3.0, 3.0, 3.0, 4.0, 0.0], "cash" => [0.0, 0.0, 0.0, 0.0, 0.0, -140.4561293636026, -140.4561293636026, -140.4561293636026, -140.4561293636026, -140.4561293636026  …  -265.320912818288, -265.320912818288, -265.320912818288, -401.866828586304, -401.866828586304, -401.866828586304, -401.866828586304, -401.866828586304, -533.5286447658507, 0.3370435884062317]), EuroCallOption{Stock}(Stock(AbstractFloat[57.0, 75.0, 65.0, 75.0, 53.0, 53.0, 54.0, 67.0, 57.0, 51.0  …  75.0, 80.0, 79.0, 77.0, 76.0, 76.0, 80.0, 73.0, 75.0, 76.0], "example_stock", 252, 2.2609836569660278), 60.0, 0.9206349206349209, 0.02, "", Dict{String, Dict{String, AbstractFloat}}("BlackScholes" => Dict("value" => 57.466422088564244))))
+(0.3370435884062317, Dict("example_stock" => [0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 1.0, 1.0, 1.0, 1.0  …  2.0, 2.0, 2.0, 3.0, 3.0, 3.0, 3.0, 3.0, 4.0, 0.0], "" => [0.0, 0.0, 0.0, 0.0, 0.0, 1.0, 1.0, 1.0, 1.0, 1.0  …  2.0, 2.0, 2.0, 3.0, 3.0, 3.0, 3.0, 3.0, 4.0, 0.0], "cash" => [0.0, 0.0, 0.0, 0.0, 0.0, -140.4561293636026, -140.4561293636026, -140.4561293636026, -140.4561293636026, -140.4561293636026  …  -265.320912818288, -265.320912818288, -265.320912818288, -401.866828586304, -401.866828586304, -401.866828586304, -401.866828586304, -401.866828586304, -533.5286447658507, 0.3370435884062317]), EuroCallOption{Stock{Int64, Int64, Float64}, Float64, Float64}(Stock{Int64, Int64, Float64}([57, 75, 65, 75, 53, 53, 54, 67, 57, 51  …  75, 80, 79, 77, 76, 76, 80, 73, 75, 76], "example_stock", 252, 2.2609836569660278), 60.0, 0.9206349206349209, 0.02, "", Dict("BlackScholes" => Dict("value" => 57.466422088564244))))
+
 ```
 
 ```@meta
