### 2023-02-17 #5056 @omendezmorales
Hi @jbytecode,

With the last modifications done by @mitchellpound, I conclude my review of the article, nice job! Kudos also for the performance improvement suggested by @lrnv .

Greetings.
