### 2023-02-18 #5056 @jbytecode
Dear @mitchellpound

I've just sent a pull request (https://github.com/USU-Analytics-Solution-Center/Bruno.jl/pull/75) for minor corrections and changes in both the manuscript and the bibtex file.

Please review the PR, and apply if you agree with them.

Thank you in advance.
