### 2023-02-18 #5056 @jbytecode
@mitchellpound - The paper seems better than ever. We have a few minor steps to go

The manuscript does not include ORCID values set for each single author. Please provide this information in your markdown file like

```
authors:
  - name: Mitchell Pound
    orcid: xxxx-xxxx-xxxx-xxxx
    affiliation: 1
```

for all of the coauthors.

Addition to that, please follows these steps carefully:

- [ ] Add ORCID information
- [ ] Compile the pdf here
- [ ] Please read the revisited paper pdf carefully, one more time, and correct any issue if it exists. Please perform a full grammar check.
- [ ] Create a tagged release in the software repository. Report the version here. The version should be in the form of `vx.y.z`, e.g. `v1.2.3`.
- [ ] Create a Zenodo archive of the repository in https://zenodo.org/
- [ ] The Zenodo archive (with the same name/title with the JOSS paper) should have the correct metadata, e.g. author names, **ORCID**s, the title of the archive. **Please set the archive title as the paper title.**
- [ ] Please provide the tagged release id, zenodo archive DOI, and Zenodo archive URL here.

Thank you in advance.
