### 2023-02-18 c5a7d33
I added some suggestions for things IDK how to fix
