### 2023-02-18 75164b0
Addressed Spencer's suggestions
