### 2023-02-19 95945b0
Grammar fix to paper.md
Found a few missing commas and clarified sentences.
