+++ paper/paper.md
@@ -41,7 +41,7 @@ Bruno allows users to compare different financial derivatives, hedging, and trad
 
 Another key feature of Bruno is that it has the ability to produce a distribution of maximum loss that could result from a trading or hedging strategy. This information is valuable to financial practitioners and market makers as it helps quantify the risk of a potential strategy before putting the strategy into place. Bruno also allows for comparison of different trading and hedging strategies. Creation of these distributions is facilitated by Bruno’s data generating processes. These processes include non-parametric methods, such as the stationary bootstrap [@politis1994stationary] with automatic block-length selection [@politis2004automatic][@patton2009correction] as well as parametric methods, such as log diffusion.
 
-Bruno was designed to be used by finance professionals and academics alike. Financial analysis of trading and hedging strategies can be intensive. This package is designed to make this type of investigation more straightforward and accessible. Many other software packages can calculate derivative prices, simulate hedging, and generate data. For example, in the Julia programing language, FinancialDerivatives.jl [@financialderivativesjl], FinancialMonteCarlo.jl [@financialmontecarlojl], and Strategems.jl [@strategemsjl] are packages that are used for derivative asset pricing, data simulation, and strategy testing, respectively. However, none of these packages have been compiled in a manner that allows for integrated analysis. Each of the listed packages performs one part of the process independently and must be assembled by the programmer. Bruno on the other hand is novel because it provides a replacement for these independent packages with a fully integrated set of tools for derivatives analysis designed to work in a unified manner. Bruno was recently used in a conference publication[@pound_2022], with several other publications nearing completion.
+Bruno was designed to be used by finance professionals and academics alike. Financial analysis of trading and hedging strategies can be intensive. This package is designed to make this type of investigation more straightforward and accessible. Many other software packages can calculate derivative prices, simulate hedging, and generate data. For example, in the Julia programing language, FinancialDerivatives.jl [@financialderivativesjl], FinancialMonteCarlo.jl [@financialmontecarlojl], and Strategems.jl [@strategemsjl] are packages that are used for derivative asset pricing, data simulation, and strategy testing, respectively. However, none of these packages have been compiled in a manner that allows for integrated analysis. Each of the listed packages performs one part of the process independently and must be assembled by the programmer. Bruno, on the other hand, is novel because it provides a replacement for these independent packages with a fully integrated set of tools for derivatives analysis designed to work in a unified manner. Bruno was recently used in a conference publication[@pound_2022], with several other publications nearing completion.
 
 # Example usage
 
@@ -78,7 +78,7 @@ end
 
 Using the type system for derivatives assets in Bruno, we define a stock and a European call option as example assets to be used in the strategy. We then run the strategy on simulated data from a log diffusion model. In this example, the European stock option is priced using the Black Scholes Model. It is important to note that alternative pricing and data simulation models could be used simply by changing the types used. This means strategies can be analyzed using a variety of assumptions about the asset and market conditions. 
 
-All logic for interest accrued and transaction costs during the time steps are all handled by the simulation environment. The code returns the cumulative return from the simulated strategy as well as the agent's holdings in the agent's portfolio at each timestep during the simulation. This allows for more complicated strategies such as those that depend on the derivative or the underlying asset, the holdings can be analyzed using common statistical time series tools. 
+All logic for interest accrued and transaction costs during the time steps are handled by the simulation environment. The code returns the cumulative return from the simulated strategy as well as the agent's holdings in the agent's portfolio at each timestep during the simulation. Thus, for more complicated strategies such as those that depend on the derivative or the underlying asset, the holdings can be analyzed using common statistical time series tools. 
 
 ```julia
 # create a random array to act as historic prices
