### 2023-02-19 #5056 @mitchellpound
@jbytecode, sorry for the delay.
The tagged release is v0.1.2
The Zenodo DOI is 10.5281/zenodo.7654894
Lastly, the Zenodo archive URL is https://zenodo.org/record/7654894#.Y_JXu9LMKV4
