### 2023-02-20 #5056 @arfon
@mitchellpound – please merge this PR which fixes a few minor issues with your references: https://github.com/USU-Analytics-Solution-Center/Bruno.jl/pull/77
