### 2023-02-20 #5056 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
